Module parachain_staking::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Bond

Collator2
Collator state with commission fee, bonded stake, and nominations

CollatorSnapshot
Snapshot of collator state at the start of the round for which they are selected

ExitQ
Store and process all delayed exits by collators and nominators

GenesisConfig
Can be used to configure the genesis state of this pallet.

Nominator
DEPRECATED nominator state

Nominator2
Nominator state

Pallet
Pallet for parachain staking

ParachainBondConfig
Reserve information { account, percent_of_inflation }

RoundInfo
The current round index and transition information

_GeneratedPrefixForStorageAtStake

_GeneratedPrefixForStorageAwardedPts

_GeneratedPrefixForStorageInflationConfig

_GeneratedPrefixForStoragePoints

_GeneratedPrefixForStorageStaked

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

CollatorStatus
The activity status of the collator

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

NominatorAdded
Convey relevant information describing if a nominator was added to the top or bottom Nominations added to the top yield a new total

NominatorStatus

Traits

Config
Configuration trait of this pallet.

Type Definitions

AtStake
Snapshot of collator nomination stake at the start of the round

AwardedPts
Points for each collator per round

BalanceOf

InflationConfig
Inflation configuration

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

Points
Total points awarded to collators for block production in the round

Staked
Total backing stake for selected candidates in the round