Module parachain_staking::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

cancel_leave_candidates

Revert the previously requested exit of the network of a collator candidate. On success, adds back the candidate to the TopCandidates and updates the collators.

candidate_stake_less

Stake less funds for a collator candidate.

candidate_stake_more

Stake more funds for a collator candidate.

claim_rewards

Claim block authoring rewards for the target address.

delegator_stake_less

Reduce the stake for delegating a collator candidate.

delegator_stake_more

Increase the stake for delegating a collator candidate.

execute_leave_candidates

Execute the network exit of a candidate who requested to leave at least ExitQueueDelay rounds ago. Prepares unstaking of the candidates and their delegators stake which can be unfreezed via unlock_unstaked after waiting at least StakeDuration many blocks.

execute_scheduled_reward_change

Executes the annual reduction of the reward rates for collators and delegators.

force_new_round

Forces the start of the new round in the next block.

force_remove_candidate

Forcedly removes a collator candidate from the TopCandidates and clears all associated storage for the candidate and their delegators.

increment_collator_rewards

Actively increment the rewards of a collator.

increment_delegator_rewards

Actively increment the rewards of a delegator.

init_leave_candidates

Request to leave the set of collator candidates.

join_candidates

Join the set of collator candidates.

join_delegators

Join the set of delegators by delegating to a collator candidate.

leave_delegators

Leave the set of delegators and, by implication, revoke the ongoing delegation.

set_blocks_per_round

Set the number of blocks each validation round lasts.

set_inflation

Set the annual inflation rate to derive per-round inflation.

set_max_candidate_stake

Set the maximal amount a collator can stake. Existing stakes are not changed.

set_max_selected_candidates

Set the maximum number of collator candidates that can be selected at the beginning of each validation round.

unlock_unstaked

Unlock all previously staked funds that are now available for unlocking by the origin account after StakeDuration blocks have elapsed.