Enum parachain_staking::pallet::Call

pub enum Call<T: Config> {

    hotfix_remove_delegation_requests {
        delegators: Vec<T::AccountId>,
    },
    hotfix_update_candidate_pool_value {
        candidates: Vec<T::AccountId>,
    },
    set_staking_expectations {
        expectations: Range<BalanceOf<T>>,
    },
    set_inflation {
        schedule: Range<Perbill>,
    },
    set_parachain_bond_account {
        new: T::AccountId,
    },
    set_parachain_bond_reserve_percent {
        new: Percent,
    },
    set_total_selected {
        new: u32,
    },
    set_collator_commission {
        new: Perbill,
    },
    set_blocks_per_round {
        new: u32,
    },
    join_candidates {
        bond: BalanceOf<T>,
        candidate_count: u32,
    },
    schedule_leave_candidates {
        candidate_count: u32,
    },
    execute_leave_candidates {
        candidate: T::AccountId,
        candidate_delegation_count: u32,
    },
    cancel_leave_candidates {
        candidate_count: u32,
    },
    go_offline {},
    go_online {},
    candidate_bond_more {
        more: BalanceOf<T>,
    },
    schedule_candidate_bond_less {
        less: BalanceOf<T>,
    },
    execute_candidate_bond_less {
        candidate: T::AccountId,
    },
    cancel_candidate_bond_less {},
    delegate {
        candidate: T::AccountId,
        amount: BalanceOf<T>,
        candidate_delegation_count: u32,
        delegation_count: u32,
    },
    schedule_leave_delegators {},
    execute_leave_delegators {
        delegator: T::AccountId,
        delegation_count: u32,
    },
    cancel_leave_delegators {},
    schedule_revoke_delegation {
        collator: T::AccountId,
    },
    delegator_bond_more {
        candidate: T::AccountId,
        more: BalanceOf<T>,
    },
    schedule_delegator_bond_less {
        candidate: T::AccountId,
        less: BalanceOf<T>,
    },
    execute_delegation_request {
        delegator: T::AccountId,
        candidate: T::AccountId,
    },
    cancel_delegation_request {
        candidate: T::AccountId,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

hotfix_remove_delegation_requests

Hotfix patch to remove all delegation requests not removed during a candidate exit

Fields of hotfix_remove_delegation_requests

delegators: Vec<T::AccountId>

hotfix_update_candidate_pool_value

Hotfix patch to correct and update CandidatePool value for candidates that have called candidate_bond_more when it did not update the CandidatePool value

Fields of hotfix_update_candidate_pool_value

candidates: Vec<T::AccountId>

set_staking_expectations

Set the expectations for total staked. These expectations determine the issuance for the round according to logic in fn compute_issuance

Fields of set_staking_expectations

expectations: Range<BalanceOf<T>>

set_inflation

Set the annual inflation rate to derive per-round inflation

Fields of set_inflation

schedule: Range<Perbill>

set_parachain_bond_account

Set the account that will hold funds set aside for parachain bond

Fields of set_parachain_bond_account

new: T::AccountId

set_parachain_bond_reserve_percent

Set the percent of inflation set aside for parachain bond

Fields of set_parachain_bond_reserve_percent

new: Percent

set_total_selected

Set the total number of collator candidates selected per round

changes are not applied until the start of the next round

Fields of set_total_selected

new: u32

set_collator_commission

Set the commission for all collators

Fields of set_collator_commission

new: Perbill

set_blocks_per_round

Set blocks per round

if called with new less than length of current round, will transition immediately in the next block

also updates per-round inflation config

Fields of set_blocks_per_round

new: u32

join_candidates

Join the set of collator candidates

Fields of join_candidates

bond: BalanceOf<T>candidate_count: u32

schedule_leave_candidates

Request to leave the set of candidates. If successful, the account is immediately removed from the candidate pool to prevent selection as a collator.

Fields of schedule_leave_candidates

candidate_count: u32

execute_leave_candidates

Execute leave candidates request

Fields of execute_leave_candidates

candidate: T::AccountIdcandidate_delegation_count: u32

cancel_leave_candidates

Cancel open request to leave candidates

only callable by collator account

result upon successful call is the candidate is active in the candidate pool

Fields of cancel_leave_candidates

candidate_count: u32

go_offline

Temporarily leave the set of collator candidates without unbonding

Fields of go_offline

go_online

Rejoin the set of collator candidates if previously had called go_offline

Fields of go_online

candidate_bond_more

Increase collator candidate self bond by more

Fields of candidate_bond_more

more: BalanceOf<T>

schedule_candidate_bond_less

Request by collator candidate to decrease self bond by less

Fields of schedule_candidate_bond_less

less: BalanceOf<T>

execute_candidate_bond_less

Execute pending request to adjust the collator candidate self bond

Fields of execute_candidate_bond_less

candidate: T::AccountId

cancel_candidate_bond_less

Cancel pending request to adjust the collator candidate self bond

Fields of cancel_candidate_bond_less

delegate

If caller is not a delegator and not a collator, then join the set of delegators If caller is a delegator, then makes delegation to change their delegation state

Fields of delegate

candidate: T::AccountIdamount: BalanceOf<T>candidate_delegation_count: u32delegation_count: u32

schedule_leave_delegators

Request to leave the set of delegators. If successful, the caller is scheduled to be allowed to exit. Success forbids future delegator actions until the request is invoked or cancelled.

Fields of schedule_leave_delegators

execute_leave_delegators

Execute the right to exit the set of delegators and revoke all ongoing delegations.

Fields of execute_leave_delegators

delegator: T::AccountIddelegation_count: u32

cancel_leave_delegators

Cancel a pending request to exit the set of delegators. Success clears the pending exit request (thereby resetting the delay upon another leave_delegators call).

Fields of cancel_leave_delegators

schedule_revoke_delegation

Request to revoke an existing delegation. If successful, the delegation is scheduled to be allowed to be revoked via the execute_delegation_request extrinsic.

Fields of schedule_revoke_delegation

collator: T::AccountId

delegator_bond_more

Bond more for delegators wrt a specific collator candidate.

Fields of delegator_bond_more

candidate: T::AccountIdmore: BalanceOf<T>

schedule_delegator_bond_less

Request bond less for delegators wrt a specific collator candidate.

Fields of schedule_delegator_bond_less

candidate: T::AccountIdless: BalanceOf<T>

execute_delegation_request

Execute pending request to change an existing delegation

Fields of execute_delegation_request

delegator: T::AccountIdcandidate: T::AccountId

cancel_delegation_request

Cancel request to change an existing delegation.

Fields of cancel_delegation_request

candidate: T::AccountId

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_hotfix_remove_delegation_requests(
delegators: Vec<T::AccountId>
) -> Self

Create a call with the variant hotfix_remove_delegation_requests.

pub fn new_call_variant_hotfix_update_candidate_pool_value(
candidates: Vec<T::AccountId>
) -> Self

Create a call with the variant hotfix_update_candidate_pool_value.

pub fn new_call_variant_set_staking_expectations(
expectations: Range<BalanceOf<T>>
) -> Self

Create a call with the variant set_staking_expectations.

pub fn new_call_variant_set_inflation(schedule: Range<Perbill>) -> Self

Create a call with the variant set_inflation.

pub fn new_call_variant_set_parachain_bond_account(new: T::AccountId) -> Self

Create a call with the variant set_parachain_bond_account.

pub fn new_call_variant_set_parachain_bond_reserve_percent(new: Percent) -> Self

Create a call with the variant set_parachain_bond_reserve_percent.

pub fn new_call_variant_set_total_selected(new: u32) -> Self

Create a call with the variant set_total_selected.

pub fn new_call_variant_set_collator_commission(new: Perbill) -> Self

Create a call with the variant set_collator_commission.

pub fn new_call_variant_set_blocks_per_round(new: u32) -> Self

Create a call with the variant set_blocks_per_round.

pub fn new_call_variant_join_candidates(
bond: BalanceOf<T>,
candidate_count: u32
) -> Self

Create a call with the variant join_candidates.

pub fn new_call_variant_schedule_leave_candidates(candidate_count: u32) -> Self

Create a call with the variant schedule_leave_candidates.

pub fn new_call_variant_execute_leave_candidates(
candidate: T::AccountId,
candidate_delegation_count: u32
) -> Self

Create a call with the variant execute_leave_candidates.

pub fn new_call_variant_cancel_leave_candidates(candidate_count: u32) -> Self

Create a call with the variant cancel_leave_candidates.

pub fn new_call_variant_go_offline() -> Self

Create a call with the variant go_offline.

pub fn new_call_variant_go_online() -> Self

Create a call with the variant go_online.

pub fn new_call_variant_candidate_bond_more(more: BalanceOf<T>) -> Self

Create a call with the variant candidate_bond_more.

pub fn new_call_variant_schedule_candidate_bond_less(less: BalanceOf<T>) -> Self

Create a call with the variant schedule_candidate_bond_less.

pub fn new_call_variant_execute_candidate_bond_less(
candidate: T::AccountId
) -> Self

Create a call with the variant execute_candidate_bond_less.

pub fn new_call_variant_cancel_candidate_bond_less() -> Self

Create a call with the variant cancel_candidate_bond_less.

pub fn new_call_variant_delegate(
candidate: T::AccountId,
amount: BalanceOf<T>,
candidate_delegation_count: u32,
delegation_count: u32
) -> Self

Create a call with the variant delegate.

pub fn new_call_variant_schedule_leave_delegators() -> Self

Create a call with the variant schedule_leave_delegators.

pub fn new_call_variant_execute_leave_delegators(
delegator: T::AccountId,
delegation_count: u32
) -> Self

Create a call with the variant execute_leave_delegators.

pub fn new_call_variant_cancel_leave_delegators() -> Self

Create a call with the variant cancel_leave_delegators.

pub fn new_call_variant_schedule_revoke_delegation(
collator: T::AccountId
) -> Self

Create a call with the variant schedule_revoke_delegation.

pub fn new_call_variant_delegator_bond_more(
candidate: T::AccountId,
more: BalanceOf<T>
) -> Self

Create a call with the variant delegator_bond_more.

pub fn new_call_variant_schedule_delegator_bond_less(
candidate: T::AccountId,
less: BalanceOf<T>
) -> Self

Create a call with the variant schedule_delegator_bond_less.

pub fn new_call_variant_execute_delegation_request(
delegator: T::AccountId,
candidate: T::AccountId
) -> Self

Create a call with the variant execute_delegation_request.

pub fn new_call_variant_cancel_delegation_request(
candidate: T::AccountId
) -> Self

Create a call with the variant cancel_delegation_request.