Struct parachain_staking::pallet::Pallet

pub struct Pallet<T>(_);

Pallet for parachain staking

Implementations

impl<T: Config> Pallet<T>

pub fn hotfix_remove_delegation_requests(
origin: OriginFor<T>,
delegators: Vec<T::AccountId>
) -> DispatchResultWithPostInfo

Hotfix patch to remove all delegation requests not removed during a candidate exit

pub fn hotfix_update_candidate_pool_value(
origin: OriginFor<T>,
candidates: Vec<T::AccountId>
) -> DispatchResultWithPostInfo

Hotfix patch to correct and update CandidatePool value for candidates that have called candidate_bond_more when it did not update the CandidatePool value

pub fn set_staking_expectations(
origin: OriginFor<T>,
expectations: Range<BalanceOf<T>>
) -> DispatchResultWithPostInfo

Set the expectations for total staked. These expectations determine the issuance for the round according to logic in fn compute_issuance

pub fn set_inflation(
origin: OriginFor<T>,
schedule: Range<Perbill>
) -> DispatchResultWithPostInfo

Set the annual inflation rate to derive per-round inflation

pub fn set_parachain_bond_account(
origin: OriginFor<T>,
new: T::AccountId
) -> DispatchResultWithPostInfo

Set the account that will hold funds set aside for parachain bond

pub fn set_parachain_bond_reserve_percent(
origin: OriginFor<T>,
new: Percent
) -> DispatchResultWithPostInfo

Set the percent of inflation set aside for parachain bond

pub fn set_total_selected(
origin: OriginFor<T>,
new: u32
) -> DispatchResultWithPostInfo

Set the total number of collator candidates selected per round

changes are not applied until the start of the next round

pub fn set_collator_commission(
origin: OriginFor<T>,
new: Perbill
) -> DispatchResultWithPostInfo

Set the commission for all collators

pub fn set_blocks_per_round(
origin: OriginFor<T>,
new: u32
) -> DispatchResultWithPostInfo

Set blocks per round

if called with new less than length of current round, will transition immediately in the next block

also updates per-round inflation config

pub fn join_candidates(
origin: OriginFor<T>,
bond: BalanceOf<T>,
candidate_count: u32
) -> DispatchResultWithPostInfo

Join the set of collator candidates

pub fn schedule_leave_candidates(
origin: OriginFor<T>,
candidate_count: u32
) -> DispatchResultWithPostInfo

Request to leave the set of candidates. If successful, the account is immediately removed from the candidate pool to prevent selection as a collator.

pub fn execute_leave_candidates(
origin: OriginFor<T>,
candidate: T::AccountId,
candidate_delegation_count: u32
) -> DispatchResultWithPostInfo

Execute leave candidates request

pub fn cancel_leave_candidates(
origin: OriginFor<T>,
candidate_count: u32
) -> DispatchResultWithPostInfo

Cancel open request to leave candidates

only callable by collator account

result upon successful call is the candidate is active in the candidate pool

pub fn go_offline(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Temporarily leave the set of collator candidates without unbonding

pub fn go_online(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Rejoin the set of collator candidates if previously had called go_offline

pub fn candidate_bond_more(
origin: OriginFor<T>,
more: BalanceOf<T>
) -> DispatchResultWithPostInfo

Increase collator candidate self bond by more

pub fn schedule_candidate_bond_less(
origin: OriginFor<T>,
less: BalanceOf<T>
) -> DispatchResultWithPostInfo

Request by collator candidate to decrease self bond by less

pub fn execute_candidate_bond_less(
origin: OriginFor<T>,
candidate: T::AccountId
) -> DispatchResultWithPostInfo

Execute pending request to adjust the collator candidate self bond

pub fn cancel_candidate_bond_less(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Cancel pending request to adjust the collator candidate self bond

pub fn delegate(
origin: OriginFor<T>,
candidate: T::AccountId,
amount: BalanceOf<T>,
candidate_delegation_count: u32,
delegation_count: u32
) -> DispatchResultWithPostInfo

If caller is not a delegator and not a collator, then join the set of delegators If caller is a delegator, then makes delegation to change their delegation state

pub fn schedule_leave_delegators(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Request to leave the set of delegators. If successful, the caller is scheduled to be allowed to exit. Success forbids future delegator actions until the request is invoked or cancelled.

pub fn execute_leave_delegators(
origin: OriginFor<T>,
delegator: T::AccountId,
delegation_count: u32
) -> DispatchResultWithPostInfo

Execute the right to exit the set of delegators and revoke all ongoing delegations.

pub fn cancel_leave_delegators(
origin: OriginFor<T>
) -> DispatchResultWithPostInfo

Cancel a pending request to exit the set of delegators. Success clears the pending exit request (thereby resetting the delay upon another leave_delegators call).

pub fn schedule_revoke_delegation(
origin: OriginFor<T>,
collator: T::AccountId
) -> DispatchResultWithPostInfo

Request to revoke an existing delegation. If successful, the delegation is scheduled to be allowed to be revoked via the execute_delegation_request extrinsic.

pub fn delegator_bond_more(
origin: OriginFor<T>,
candidate: T::AccountId,
more: BalanceOf<T>
) -> DispatchResultWithPostInfo

Bond more for delegators wrt a specific collator candidate.

pub fn schedule_delegator_bond_less(
origin: OriginFor<T>,
candidate: T::AccountId,
less: BalanceOf<T>
) -> DispatchResultWithPostInfo

Request bond less for delegators wrt a specific collator candidate.

pub fn execute_delegation_request(
origin: OriginFor<T>,
delegator: T::AccountId,
candidate: T::AccountId
) -> DispatchResultWithPostInfo

Execute pending request to change an existing delegation

pub fn cancel_delegation_request(
origin: OriginFor<T>,
candidate: T::AccountId
) -> DispatchResultWithPostInfo

Cancel request to change an existing delegation.

impl<T: Config> Pallet<T>

pub fn is_delegator(acc: &T::AccountId) -> bool

pub fn is_candidate(acc: &T::AccountId) -> bool

pub fn is_selected_candidate(acc: &T::AccountId) -> bool

pub fn compute_top_candidates() -> Vec<T::AccountId>

Compute the top TotalSelected candidates in the CandidatePool and return a vec of their AccountIds (in the order of selection)

impl<T: Config> Pallet<T>

pub fn error_metadata() -> Option<PalletErrorMetadata>

impl<T: Config> Pallet<T>

pub fn collator_commission() -> Perbill

Commission percent taken off of rewards for all collators

impl<T: Config> Pallet<T>

pub fn total_selected() -> u32

The total candidates selected every round

impl<T: Config> Pallet<T>

pub fn parachain_bond_info() -> ParachainBondConfig<T::AccountId>

Parachain bond config info { account, percent_of_inflation }

impl<T: Config> Pallet<T>

pub fn round() -> RoundInfo<T::BlockNumber>

Current round index and next round scheduled transition

impl<T: Config> Pallet<T>

pub fn nominator_state2<KArg>(
k: KArg
) -> Option<Nominator2<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

DEPRECATED in favor of DelegatorState Get nominator state associated with an account if account is nominating else None

impl<T: Config> Pallet<T>

pub fn delegator_state<KArg>(
k: KArg
) -> Option<Delegator<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Get delegator state associated with an account if account is delegating else None

impl<T: Config> Pallet<T>

pub fn candidate_state<KArg>(
k: KArg
) -> Option<CollatorCandidate<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

DEPRECATED Get collator candidate state associated with an account if account is a candidate else None

impl<T: Config> Pallet<T>

pub fn candidate_info<KArg>(k: KArg) -> Option<CandidateMetadata<BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Get collator candidate info associated with an account if account is candidate else None

impl<T: Config> Pallet<T>

pub fn top_delegations<KArg>(
k: KArg
) -> Option<Delegations<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Top delegations for collator candidate

impl<T: Config> Pallet<T>

pub fn bottom_delegations<KArg>(
k: KArg
) -> Option<Delegations<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Bottom delegations for collator candidate

impl<T: Config> Pallet<T>

pub fn collator_state2<KArg>(
k: KArg
) -> Option<Collator2<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

DEPRECATED in favor of CandidateState Get collator state associated with an account if account is collating else None

impl<T: Config> Pallet<T>

pub fn selected_candidates() -> Vec<T::AccountId>

The collator candidates selected for the current round

impl<T: Config> Pallet<T>

pub fn total() -> BalanceOf<T>

Total capital locked by this staking pallet

impl<T: Config> Pallet<T>

pub fn candidate_pool() -> OrderedSet<Bond<T::AccountId, BalanceOf<T>>>

The pool of collator candidates, each with their total backing stake

impl<T: Config> Pallet<T>

pub fn at_stake<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> CollatorSnapshot<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<u32>,
KArg2: EncodeLike<T::AccountId>,

Snapshot of collator delegation stake at the start of the round

impl<T: Config> Pallet<T>

pub fn delayed_payouts<KArg>(k: KArg) -> Option<DelayedPayout<BalanceOf<T>>> where
KArg: EncodeLike<u32>,

Delayed payouts

impl<T: Config> Pallet<T>

pub fn staked<KArg>(k: KArg) -> BalanceOf<T> where
KArg: EncodeLike<u32>,

Total counted stake for selected candidates in the round

impl<T: Config> Pallet<T>

pub fn inflation_config() -> InflationInfo<BalanceOf<T>>

Inflation configuration

impl<T: Config> Pallet<T>

pub fn points<KArg>(k: KArg) -> u32 where
KArg: EncodeLike<u32>,

Total points awarded to collators for block production in the round

impl<T: Config> Pallet<T>

pub fn awarded_pts<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> u32 where
KArg1: EncodeLike<u32>,
KArg2: EncodeLike<T::AccountId>,

Points for each collator per round