Module parachain_staking::types

Types for parachain-staking

Structs

Bond

CandidateBondLessRequest

Request scheduled to change the collator candidate self-bond

CandidateMetadata

All candidate info except the top and bottom delegations

Collator2

DEPRECATED Collator state with commission fee, bonded stake, and delegations

CollatorCandidate

DEPRECATED, replaced by CandidateMetadata and two storage instances of Delegations Collator candidate state with self bond + delegations

CollatorSnapshot

Snapshot of collator state at the start of the round for which they are selected

DelayedPayout

Info needed to make delayed payments to stakers after round end

DelegationRequest

Delegations

Type for top and bottom delegation storage item

Delegator

Delegator state

Nominator2

DEPRECATED in favor of Delegator Nominator state

ParachainBondConfig

Reserve information { account, percent_of_inflation }

PendingDelegationRequests

Pending requests to mutate delegations for each delegator

RoundInfo

The current round index and transition information

Enums

CapacityStatus

Capacity status for top or bottom delegations

CollatorStatus

The activity status of the collator

DelegationChange

Changes requested by the delegator

DelegatorAdded

Convey relevant information describing if a delegator was added to the top or bottom Delegations added to the top yield a new total

DelegatorStatus