Crate parachain_staking

Parachain Staking

Minimal staking pallet that implements collator selection by total backed stake. The main difference between this pallet and frame/pallet-staking is that this pallet uses direct delegation. Nominators choose exactly who they nominate and with what stake. This is different from frame/pallet-staking where you approval vote and then run Phragmen.

Rules

There is a new round every <Round<T>>::get().length blocks.

At the start of every round,

issuance is distributed to collators (and their nominators) for block authoring T::RewardPaymentDelay rounds ago

queued collator and nominator exits are executed

a new set of collators is chosen from the candidates

To join the set of candidates, call join_candidates with bond >= MinCollatorCandidateStk.

To leave the set of candidates, call leave_candidates. If the call succeeds, the collator is removed from the pool of candidates so they cannot be selected for future collator sets, but they are not unstaked until T::LeaveCandidatesDelay rounds later. The exit request is stored in the ExitQueue and processed T::LeaveCandidatesDelay rounds later to unstake the collator and all of its nominations.

To join the set of nominators, call nominate and pass in an account that is already a collator candidate and bond >= MinNominatorStk. Each nominator can nominate up to T::MaxCollatorsPerNominator collator candidates by calling nominate.

To revoke a nomination, call revoke_nomination with the collator candidate’s account. To leave the set of nominators and revoke all nominations, call leave_nominators.

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights
Autogenerated weights for parachain_staking

Structs

InflationInfo

Range