Crate parachain_staking

Parachain Staking

Minimal staking pallet that implements collator selection by total backed stake. The main difference between this pallet and frame/pallet-staking is that this pallet uses direct delegation. Nominators choose exactly who they nominate and with what stake. This is different from frame/pallet-staking where you approval vote and then run Phragmen.

Rules

There is a new round every BlocksPerRound blocks.

At the start of every round,

issuance is distributed to collators for BondDuration rounds ago in proportion to the points they received in that round (for authoring blocks)

queued collator exits are executed

a new set of collators is chosen from the candidates

To join the set of candidates, an account must call join_candidates with stake >= MinCollatorCandidateStk and fee <= MaxFee. The fee is taken off the top of any rewards for the collator before the remaining rewards are distributed in proportion to stake to all nominators (including the collator, who always self-nominates).

To leave the set of candidates, the collator calls leave_candidates. If the call succeeds, the collator is removed from the pool of candidates so they cannot be selected for future collator sets, but they are not unstaked until BondDuration rounds later. The exit request is stored in the ExitQueue and processed BondDuration rounds later to unstake the collator and all of its nominators.

To join the set of nominators, an account must call join_nominators with stake >= MinNominatorStk. There are also runtime methods for nominating additional collators and revoking nominations.

Re-exports

pub use pallet::*;

Modules

pallet

Structs

InflationInfo

Range