Module parachain_staking::pallet

Structs

Bond

Collator
Global collator state with commission fee, bonded stake, and nominations

CollatorSnapshot
Snapshot of collator state at the start of the round for which they are selected

GenesisConfig

Nominator

Pallet
Pallet for parachain staking

_GeneratedPrefixForStorageAtStake

_GeneratedPrefixForStorageAwardedPts

_GeneratedPrefixForStorageInflationConfig

_GeneratedPrefixForStoragePoints

_GeneratedPrefixForStorageStaked

Enums

Call

CollatorStatus
The activity status of the collator

Error

Event

Traits

Config
Configuration trait of this pallet.

Type Definitions

AtStake
Snapshot of collator nomination stake at the start of the round

AwardedPts
Points for each collator per round

BalanceOf

InflationConfig
Inflation configuration

Module
Type alias to Pallet, to be used by construct_runtime.

Points
Total points awarded to collators for block production in the round

Staked
Total backing stake for selected candidates in the round