Module parachain_staking::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

Pallet for parachain staking

_GeneratedPrefixForStorageAtStake

_GeneratedPrefixForStorageAwardedPts

_GeneratedPrefixForStorageDelayedPayouts

_GeneratedPrefixForStorageInflationConfig

_GeneratedPrefixForStoragePoints

_GeneratedPrefixForStorageStaked

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

AtStake

Snapshot of collator delegation stake at the start of the round

AwardedPts

Points for each collator per round

BalanceOf

DelayedPayouts

Delayed payouts

InflationConfig

Inflation configuration

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Points

Total points awarded to collators for block production in the round

Staked

Total counted stake for selected candidates in the round