Trait parachain_staking::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type MonetaryGovernanceOrigin: EnsureOrigin<Self::Origin>;
    type MinBlocksPerRound: Get<u32>;
    type DefaultBlocksPerRound: Get<u32>;
    type LeaveCandidatesDelay: Get<u32>;
    type LeaveNominatorsDelay: Get<u32>;
    type RevokeNominationDelay: Get<u32>;
    type RewardPaymentDelay: Get<u32>;
    type MinSelectedCandidates: Get<u32>;
    type MaxNominatorsPerCollator: Get<u32>;
    type MaxCollatorsPerNominator: Get<u32>;
    type DefaultCollatorCommission: Get<Perbill>;
    type DefaultParachainBondReservePercent: Get<Percent>;
    type MinCollatorStk: Get<BalanceOf<Self>>;
    type MinCollatorCandidateStk: Get<BalanceOf<Self>>;
    type MinNomination: Get<BalanceOf<Self>>;
    type MinNominatorStk: Get<BalanceOf<Self>>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The currency type

type MonetaryGovernanceOrigin: EnsureOrigin<Self::Origin>

The origin for monetary governance

type MinBlocksPerRound: Get<u32>

Minimum number of blocks per round

type DefaultBlocksPerRound: Get<u32>

Default number of blocks per round at genesis

type LeaveCandidatesDelay: Get<u32>

Number of rounds that collators remain bonded before exit request is executed

type LeaveNominatorsDelay: Get<u32>

Number of rounds that nominators remain bonded before exit request is executed

type RevokeNominationDelay: Get<u32>

Number of rounds that nominations remain bonded before revocation request is executed

type RewardPaymentDelay: Get<u32>

Number of rounds after which block authors are rewarded

type MinSelectedCandidates: Get<u32>

Minimum number of selected candidates every round

type MaxNominatorsPerCollator: Get<u32>

Maximum nominators counted per collator

type MaxCollatorsPerNominator: Get<u32>

Maximum collators per nominator

type DefaultCollatorCommission: Get<Perbill>

Default commission due to collators, set at genesis

type DefaultParachainBondReservePercent: Get<Percent>

Default percent of inflation set aside for parachain bond account

type MinCollatorStk: Get<BalanceOf<Self>>

Minimum stake required for any account to be in SelectedCandidates for the round

type MinCollatorCandidateStk: Get<BalanceOf<Self>>

Minimum stake required for any account to be a collator candidate

type MinNomination: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to nominate

type MinNominatorStk: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to become a nominator

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.