Trait parachain_staking::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type BlocksPerRound: Get<u32>;
    type BondDuration: Get<u32>;
    type TotalSelectedCandidates: Get<u32>;
    type MaxNominatorsPerCollator: Get<u32>;
    type MaxCollatorsPerNominator: Get<u32>;
    type MaxFee: Get<Perbill>;
    type MinCollatorStk: Get<BalanceOf<Self>>;
    type MinCollatorCandidateStk: Get<BalanceOf<Self>>;
    type MinNomination: Get<BalanceOf<Self>>;
    type MinNominatorStk: Get<BalanceOf<Self>>;
}

Configuration trait of this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

Overarching event type

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The currency type

type BlocksPerRound: Get<u32>

Number of blocks per round

type BondDuration: Get<u32>

Number of rounds that collators remain bonded before exit request is executed

type TotalSelectedCandidates: Get<u32>

Total number of selected candidates every round

type MaxNominatorsPerCollator: Get<u32>

Maximum nominators per collator

type MaxCollatorsPerNominator: Get<u32>

Maximum collators per nominator

type MaxFee: Get<Perbill>

Maximum fee for any collator

type MinCollatorStk: Get<BalanceOf<Self>>

Minimum stake required for any account to be in SelectedCandidates for the round

type MinCollatorCandidateStk: Get<BalanceOf<Self>>

Minimum stake required for any account to be a collator candidate

type MinNomination: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to nominate

type MinNominatorStk: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to become a nominator

Loading content...