Trait parachain_staking::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type MonetaryGovernanceOrigin: EnsureOrigin<Self::Origin>;
    type MinBlocksPerRound: Get<u32>;
    type DefaultBlocksPerRound: Get<u32>;
    type LeaveCandidatesDelay: Get<u32>;
    type CandidateBondLessDelay: Get<u32>;
    type LeaveDelegatorsDelay: Get<u32>;
    type RevokeDelegationDelay: Get<u32>;
    type DelegationBondLessDelay: Get<u32>;
    type RewardPaymentDelay: Get<u32>;
    type MinSelectedCandidates: Get<u32>;
    type MaxTopDelegationsPerCandidate: Get<u32>;
    type MaxBottomDelegationsPerCandidate: Get<u32>;
    type MaxDelegationsPerDelegator: Get<u32>;
    type DefaultCollatorCommission: Get<Perbill>;
    type DefaultParachainBondReservePercent: Get<Percent>;
    type MinCollatorStk: Get<BalanceOf<Self>>;
    type MinCandidateStk: Get<BalanceOf<Self>>;
    type MinDelegation: Get<BalanceOf<Self>>;
    type MinDelegatorStk: Get<BalanceOf<Self>>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Overarching event type

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The currency type

type MonetaryGovernanceOrigin: EnsureOrigin<Self::Origin>

The origin for monetary governance

type MinBlocksPerRound: Get<u32>

Minimum number of blocks per round

type DefaultBlocksPerRound: Get<u32>

Default number of blocks per round at genesis

type LeaveCandidatesDelay: Get<u32>

Number of rounds that candidates remain bonded before exit request is executable

type CandidateBondLessDelay: Get<u32>

Number of rounds candidate requests to decrease self-bond must wait to be executable

type LeaveDelegatorsDelay: Get<u32>

Number of rounds that delegators remain bonded before exit request is executable

type RevokeDelegationDelay: Get<u32>

Number of rounds that delegations remain bonded before revocation request is executable

type DelegationBondLessDelay: Get<u32>

Number of rounds that delegation less requests must wait before executable

type RewardPaymentDelay: Get<u32>

Number of rounds after which block authors are rewarded

type MinSelectedCandidates: Get<u32>

Minimum number of selected candidates every round

type MaxTopDelegationsPerCandidate: Get<u32>

Maximum top delegations counted per candidate

type MaxBottomDelegationsPerCandidate: Get<u32>

Maximum bottom delegations (not counted) per candidate

type MaxDelegationsPerDelegator: Get<u32>

Maximum delegations per delegator

type DefaultCollatorCommission: Get<Perbill>

Default commission due to collators, is CollatorCommission storage value in genesis

type DefaultParachainBondReservePercent: Get<Percent>

Default percent of inflation set aside for parachain bond account

type MinCollatorStk: Get<BalanceOf<Self>>

Minimum stake required for any candidate to be in SelectedCandidates for the round

type MinCandidateStk: Get<BalanceOf<Self>>

Minimum stake required for any account to be a collator candidate

type MinDelegation: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to delegate

type MinDelegatorStk: Get<BalanceOf<Self>>

Minimum stake for any registered on-chain account to be a delegator

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.